refactor(server): move connection logic to own files
