+++ server/src/connection/mod.rs
@@ -1,0 +1,7 @@
+mod tcp;
+#[cfg(unix)]
+mod uds;
+
+pub use tcp::tcp_stream;
+#[cfg(unix)]
+pub use uds::uds_stream;
